requirements.txt
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==1.5.0
certifi==2016.9.26
colorama==0.3.7
configparser==3.5.0
decorator==4.0.10
Django==1.10.4
djangorestframework==3.5.3
entrypoints==0.2.2
enum34==1.1.6
functools32==3.2.3.post2
html5lib==0.9999999
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
Jinja2==2.8
jsonschema==2.5.1
jupyter-client==4.4.0
jupyter-core==4.2.1
Markdown==2.6.7
MarkupSafe==0.23
mistune==0.7.3
nbconvert==5.0.0
nbformat==4.2.0
notebook==4.3.1
numpy==1.11.3
pandas==0.19.1
pandocfilters==1.4.1
pathlib2==2.1.0
pickleshare==0.7.4
prompt-toolkit==1.0.9
Pygments==2.1.3
python-dateutil==2.6.0
pytz==2016.10
pyzmq==16.0.2
redis==2.10.5
selenium==2.53.6
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
testpath==0.3
tornado==4.4.2
traitlets==4.3.1
wcwidth==0.1.7
win-unicode-console==0.5
